+++ misskey-util/src/builder.rs
@@ -17,13 +17,15 @@ mod messaging;
 mod misc;
 mod note;
 mod page;
-mod user;
 
 #[cfg(feature = "12-47-0")]
 mod channel;
 
 #[cfg(feature = "12-79-0")]
 mod gallery;
+
+#[cfg(not(feature = "12-88-0"))]
+mod user;
 
 pub use admin::{
     AnnouncementUpdateBuilder, EmojiUpdateBuilder, MetaUpdateBuilder, ServerLogListBuilder,
@@ -38,7 +40,6 @@ pub use me::{IntoUserFields, MeUpdateBuilder};
 pub use messaging::MessagingMessageBuilder;
 pub use note::NoteBuilder;
 pub use page::{PageBuilder, PageUpdateBuilder};
-pub use user::UserListBuilder;
 
 #[cfg(feature = "12-47-0")]
 #[cfg_attr(docsrs, doc(cfg(feature = "12-47-0")))]
@@ -59,3 +60,8 @@ pub use gallery::GalleryPostUpdateBuilder;
 #[cfg(feature = "12-80-0")]
 #[cfg_attr(docsrs, doc(cfg(feature = "12-80-0")))]
 pub use admin::{AdBuilder, AdUpdateBuilder};
+
+// misskey-dev/misskey#7656
+#[cfg(not(feature = "12-88-0"))]
+#[cfg_attr(docsrs, doc(cfg(not(feature = "12-88-0"))))]
+pub use user::UserListBuilder;